Remove data grants from app framework
